Added main activity Added workout actvity Added calibration activity Added styles
