+++ app/src/main/java/com/example/android/madcowtest/CalibrationActivity.java
@@ -1,0 +1,13 @@
+package com.example.android.madcowtest;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class CalibrationActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_calibration);
+    }
+}
