+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">Madcow Test</string>
+    <string name="workout">Workout</string>
+    <string name="calibrate">Calibrate</string>
 </resources>
